Update blog system
This commit introduces a few radical changes to the blog app.

* Revisions have been removed. All blogpost information is stored in
  the Post object. While having revisions is a nice to have feature,
  the current implementation was getting in the way of having a
  simple and predictable publishing experience.
* Add class-based views for post-list and post-detail. The blog
  system is simple enough to rely on the default functionality
  provided by class-based views (for example pagination).
* Rename some model properties to be more consistent with other
  blogging platforms.

As a consequence of these changes, indexing has also been simplified,
since now we can directly work with Posts, avoiding references to the
latest published Revision.

In order to deploy this in production, it is necessary to run the
0012_drop_revisions migration, which will take care of merging the
latest revision of a post into the post itself.