Add background video processing
Once a Static Asset of type 'video' is added, a background job to
perform video encoding and thumbnailing is started. This is done
through the django-background-tasks app and through the Coconut video
processing service.

Video processing happens as follows:
- a job description is created and transmitted to the service
- the source video is fetched from the main bucket
- processed thumbnail and encoded video are uploaded to a staging
  bucket
- this triggers a webhook call on the app, which in turn starts a new
  job, which moves the files from the staging bucket to the main
  bucket

A 'Process videos for selected assets' command is also available in
the admin backend, to reprocess existing Static Assets if needed.