Temporarily disable "looper" and "subscriptions"
The devfund's version of looper that is soon going to be used
in Studio has a different set of migrations, incompatible
with those that exist in current Studio's looper.
All of the looper tables have to be dropped manually to make it
possible to install looper anew with this new set of migrations.
This is going to be done with the following command (before this commit
is deployed):

    python manage.py migrate looper zero

Because "subscriptions" app depends on looper, it and its tables
also have to be removed:

    python manage.py migrate subscriptions zero

After this is done, this commit can be deployed, and then
looper's version can be updated, which will result in new migrations.

Note, that neither of these apps are currently in active use
and don't have any data stored in their database tables.