Add Production Credits for users
Assign a production credit to subscribers that support a film
production for a specified period of time.

Users can check their production credits at the
/settings/production-credits/ endpoint (if they have at least one
credit), and in at /films/<film_slug>/production-credit/. In the film
specific endpoint, a user can opt-in to display their name in the
end-credits of the film.

Credits are generated by an admin through the
'create_production_credits' management command. Credits can be
'locked' by an admin to prevent future edits.
A formatted text file with the credits for a specific production can
be generated through the 'generate_production_credits_scroll' command.